backend/requirements.txt
2captcha-python==1.2.2
aniso8601==9.0.1
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
fsspec==2023.10.0
gevent==23.9.1
greenlet==3.0.1
idna==3.4
imageio==2.32.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
mysql-connector-python==8.2.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
opencv-python-headless==4.8.1.78
packaging==23.2
protobuf==4.21.12
pyclipper==1.3.0.post5
PyJWT==2.8.0
python-bidi==0.4.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
schedule==1.2.1
scikit-image==0.22.0
scipy==1.11.3
shapely==2.0.2
six==1.16.0
soupsieve==2.5
sympy==1.12
tifffile==2023.9.26
TwoCaptcha==0.0.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
